extras/requirements.txt
poster
numpy
configparser==3.5.0
pyqtgraph
pandas
matplotlib
pyserial
